pyproject.toml
[project]
name = "5langchain-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.36.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.8",
    "langchain-core>=1.2.9",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "numpy>=2.4.2",
    "openai>=2.17.0",
    "pandas>=2.3.3",
    "pip>=26.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.54.0",
    "transformers>=4.57.6",
]
